+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/beneficiaries/BeneficiaryDto.kt
@@ -18,15 +18,15 @@ import com.google.gson.annotations.SerializedName
  * @param status The status of the beneficiary, active = 1, inactive = 0.
  */
 data class BeneficiaryDto (
-    @SerializedName("idBeneficiario") val id: String,
-    @SerializedName("nombreBeneficiario") val name: String,
-    @SerializedName("fechaNacimiento") val birthdate: String,
-    @SerializedName("numeroContactoEmergencia") val emergencyNumber: String,
-    @SerializedName("nombreContactoEmergencia") val emergencyName: String,
-    @SerializedName("relacionContactoEmergencia") val emergencyRelation: String,
-    @SerializedName("descripcion") val details: String,
-    @SerializedName("fechaIngreso") val entryDate: String,
-    @SerializedName("foto") val image: String,
-    @SerializedName("discapacidad") val disabilities: String,
-    @SerializedName("estatus") val status: Int
+    @SerializedName("idBeneficiario") val id: String?,
+    @SerializedName("nombreBeneficiario") val name: String?,
+    @SerializedName("fechaNacimiento") val birthdate: String?,
+    @SerializedName("numeroContactoEmergencia") val emergencyNumber: String?,
+    @SerializedName("nombreContactoEmergencia") val emergencyName: String?,
+    @SerializedName("relacionContactoEmergencia") val emergencyRelation: String?,
+    @SerializedName("descripcion") val details: String?,
+    @SerializedName("fechaIngreso") val entryDate: String?,
+    @SerializedName("foto") val image: String?,
+    @SerializedName("discapacidad") val disabilities: String?,
+    @SerializedName("estatus") val status: Int?
 )