+++ app/src/main/java/com/app/arcabyolimpo/presentation/navigation/NavGraph.kt
@@ -29,6 +29,8 @@ import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordR
 import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordRegistrationSuccessScreen
 import com.app.arcabyolimpo.presentation.screens.splash.SplashScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplyDetail.SuppliesDetailScreen
+import com.app.arcabyolimpo.presentation.screens.beneficiary.BeneficiaryDetailScreen
+import com.app.arcabyolimpo.presentation.screens.beneficiary.BeneficiaryList
 import com.app.arcabyolimpo.presentation.screens.supply.supplyList.SupplyListScreen
 import com.app.arcabyolimpo.presentation.screens.tokenverification.TokenVerificationFailedScreen
 import com.app.arcabyolimpo.presentation.screens.tokenverification.TokenVerificationViewModel
@@ -81,12 +83,10 @@ sealed class Screen(
 
     object AddNewWorkshop : Screen("workshop/add")
 
-    object BeneficiaryList : Screen("beneficiary")
+    object BeneficiaryList : Screen("beneficiary_list")
 
-    object BeneficiaryDetail : Screen("beneficiary/id")
-
-    object SupplyDetail : Screen("supply/{idSupply}") {
-        fun createRoute(idSupply: String) = "supply/$idSupply"
+    object BeneficiaryDetail : Screen("beneficiary_detail/{beneficiaryId}") {
+        fun createRoute(beneficiaryId: String) = "beneficiary_detail/$beneficiaryId"
     }
 }
 
@@ -418,15 +418,37 @@ fun ArcaNavGraph(
                 onClickDelete = {
                     // TODO: Add when delete a supply is ready
                 },
-            onClickModify = {
-                // TODO: Add when modify a supply is ready
-            },
-            modifySupplyBatch = {
-                // TODO: Add when modify a supply batch is ready
-            },
-            deleteSupplyBatch = {
-                // TODO: Add when delete a supply batch is ready
-            },
+            
+            )
+        }
+        /**
+         * Beneficiary List Screen.
+         *
+         * Shows the grid of beneficiaries.
+         */
+        composable(Screen.BeneficiaryList.route) {
+            BeneficiaryList(
+                onBeneficiaryClick = { beneficiaryId ->
+                    navController.navigate(Screen.BeneficiaryDetail.createRoute(beneficiaryId))
+                },
+                onFilterClick = { /* TODO: Lógica de VM */},
+                onNotificationClick = { /* TODO: Lógica de VM */}
+            )
+        }
+
+        /**
+         * Beneficiary Detail Screen.
+         *
+         * Shows the details of a beneficiary and allows the function to eliminate them, others functionality are in progress.
+         */
+        composable(
+            route = Screen.BeneficiaryDetail.route,
+            arguments = listOf(navArgument("beneficiaryId") { type = NavType.StringType})
+        ) {
+            BeneficiaryDetailScreen(
+                onBackClick = { navController.popBackStack() },
+                onModifyClick = { /* TODO: Lógica de VM */ },
+                viewModel = hiltViewModel()
             )
         }
     }