+++ CHANGELOG.md
@@ -4,7 +4,8 @@
  -  Fixed Main Activity for changes made for supply
 
 ### Added
- - Consulting an specific Supply with the batches it is contained into.
+- Added final comments to the function delete beneficiaries
+- Consulting an specific Supply with the batches it is contained into.
 - Added customization of Snackbar Atom
 
 ## [0.1.0] - 10/11/2025